chore: rename model.ts to messageModel.ts to match class name
