+++ src/storage/sets/flatbuffers/signerSet.ts
@@ -1,6 +1,6 @@
 import RocksDB, { Transaction } from '~/storage/db/binaryrocksdb';
 import { BadRequestError } from '~/utils/errors';
-import MessageModel from '~/storage/flatbuffers/model';
+import MessageModel from '~/storage/flatbuffers/messageModel';
 import { ResultAsync } from 'neverthrow';
 import { SignerAddModel, UserPrefix, SignerRemoveModel } from '~/storage/flatbuffers/types';
 import { isSignerAdd, isSignerRemove } from '~/storage/flatbuffers/typeguards';
